requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
django-filter==24.1
PyJWT==2.10.1
cryptography==42.0.5
python-dotenv==1.1.1
mysqlclient==2.2.7
sqlparse==0.5.3
mimesis==13.1.0
python-dateutil==2.9.0
